added for_each_block method; migrated DumpAsm() to use this

+++ asm/asm-proto.cpp
@@ -6,6 +6,6 @@ namespace zc::z80 {
    const MemoryLocation FP_loc(&FP_register_value);
    const MemoryValue FP_memval(&FP_loc, long_size); /* NOTE: long size because it'll store
                                                      * the return address. */
-   const IndexedRegisterValue FP_idxval(&rv_ix, 0);
+   const IndexedRegisterValue FP_idxval(&rv_ix, (int8_t) 0);
   
 }
